feat: serve verb, daemon default
