+++ cmd/daytona/config/config.go
@@ -45,7 +45,7 @@ type GitProvider struct {
 	Name string
 }
 
-const configDoesntExistError = "config does not exist. Run `daytona server` to create a default profile or `daytona profile add` to connect to a remote server."
+const configDoesntExistError = "config does not exist. Run `daytona serve` to create a default profile or `daytona profile add` to connect to a remote server."
 
 func IsNotExist(err error) bool {
 	return err.Error() == configDoesntExistError
